# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "FolderCreate": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FolderCreate",
        "type": "object"
      },
      "FolderInfo": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "registered": {
            "default": false,
            "title": "Registered",
            "type": "boolean"
          },
          "root_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Id"
          },
          "selector": {
            "title": "Selector",
            "type": "string"
          }
        },
        "required": [
          "name",
          "selector",
          "count"
        ],
        "title": "FolderInfo",
        "type": "object"
      },
      "FolderRelocate": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FolderRelocate",
        "type": "object"
      },
      "FolderRelocateResult": {
        "properties": {
          "active_tool_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Tool Id"
          },
          "files_updated": {
            "title": "Files Updated",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "root_id": {
            "title": "Root Id",
            "type": "string"
          },
          "selector": {
            "title": "Selector",
            "type": "string"
          },
          "status": {
            "const": "relocated",
            "title": "Status",
            "type": "string"
          },
          "sync": {
            "title": "Sync",
            "type": "string"
          }
        },
        "required": [
          "status",
          "name",
          "selector",
          "path",
          "root_id",
          "files_updated",
          "sync"
        ],
        "title": "FolderRelocateResult",
        "type": "object"
      },
      "FolderRemovalPreview": {
        "properties": {
          "external_images_affected": {
            "default": 0,
            "title": "External Images Affected",
            "type": "integer"
          },
          "indexed_files": {
            "title": "Indexed Files",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "root_id": {
            "title": "Root Id",
            "type": "string"
          },
          "sidecar_bytes": {
            "title": "Sidecar Bytes",
            "type": "integer"
          },
          "sidecar_files": {
            "title": "Sidecar Files",
            "type": "integer"
          },
          "sidecar_history_preserved": {
            "default": true,
            "title": "Sidecar History Preserved",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "path",
          "root_id",
          "indexed_files",
          "sidecar_files",
          "sidecar_bytes"
        ],
        "title": "FolderRemovalPreview",
        "type": "object"
      },
      "FolderRemovalResult": {
        "properties": {
          "external_images_affected": {
            "default": 0,
            "title": "External Images Affected",
            "type": "integer"
          },
          "files_removed": {
            "title": "Files Removed",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "unindex_only",
              "delete_sidecars"
            ],
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sidecar_bytes_removed": {
            "default": 0,
            "title": "Sidecar Bytes Removed",
            "type": "integer"
          },
          "sidecar_files_removed": {
            "default": 0,
            "title": "Sidecar Files Removed",
            "type": "integer"
          },
          "sidecar_history_preserved": {
            "default": true,
            "title": "Sidecar History Preserved",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "name",
          "mode",
          "files_removed"
        ],
        "title": "FolderRemovalResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Keivotos - Waifu-Hoard",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/folders": {
      "get": {
        "operationId": "list_folders_api_folders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FolderInfo"
                  },
                  "title": "Response List Folders Api Folders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Folders"
      },
      "post": {
        "operationId": "register_folder_api_folders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Folder"
      }
    },
...
    "/api/folders/{folder_identifier}": {
      "delete": {
        "operationId": "remove_folder_api_folders__folder_identifier__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder_identifier",
            "required": true,
            "schema": {
              "title": "Folder Identifier",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "unindex_only",
              "enum": [
                "unindex_only",
                "delete_sidecars"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRemovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Folder"
      }
    },
    "/api/folders/{folder_identifier}/removal-preview": {
      "get": {
        "operationId": "folder_removal_preview_api_folders__folder_identifier__removal_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder_identifier",
            "required": true,
            "schema": {
              "title": "Folder Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRemovalPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Folder Removal Preview"
      }
    },
    "/api/folders/{folder_identifier}/rescan": {
      "post": {
        "operationId": "rescan_folder_api_folders__folder_identifier__rescan_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder_identifier",
            "required": true,
            "schema": {
              "title": "Folder Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rescan Folder"
      }
    },
    "/api/folders/{root_id}/path": {
      "put": {
        "description": "Point a stable root identity at a folder that the user moved on disk.",
        "operationId": "relocate_folder_api_folders__root_id__path_put",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "title": "Root Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderRelocate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRelocateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relocate Folder"
      }
    },
...
  }
}